requirements.txt
streamlit==1.22.0
tensorflow==2.12.0
numpy==1.23.5
Pillow==9.5.0
streamlit-drawable-canvas==0.9.2
